test: refactor OTLP metric validators
Move common validator logic for gRPC to
test/common/nsolid-grpc-agent/validators.js.
Add a new validator for OTLP Metrics: `checkOTLPMetricsData()`.
Add also a helper method `hasThreadAttributes()` to check whether a
metric datapoint is a thread or a process.
Refactor `test/agents/test-grpc-metrics.mjs` to use these new helpers.
